fix : 사용자 위치 소수점

+++ sharedBook/src/main/java/com/bookstore/sharedBook/book/dto/response/ShelfDetailResponseDto.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -20,8 +21,8 @@ public class ShelfDetailResponseDto {
     private String userId;
     private String status;
     private Integer token;
-    private Long latitude;
-    private Long longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private List<String> images;
 
     public static ShelfDetailResponseDto toShelfResponseDto(Shelf shelf, List<String> images){
